third_party/blink/web_tests/accessibility/long-text.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<pre id="pre" style="word-wrap: break-word; white-space: pre-wrap;"><script>
for (let count = 1; count <= 250; count ++)
  document.write('' + count + '. The quick brown fox jumps over the lazy dog. ');
</script></pre>
</body>

<script>
test(function(t) {
    var axPre = accessibilityController.accessibleElementById("pre");
    var axStaticText = axPre.childAtIndex(0);
    console.log(axStaticText.name);
    assert_equals(axStaticText.name.substr(-50), "250. The quick brown fox jumps over the lazy dog. ");
}, "Accessible name should not be truncated.");
</script>
